+++ src/Symfony/Bundle/SecurityBundle/DependencyInjection/Compiler/AddExpressionLanguageProvidersPass.php
@@ -30,10 +30,5 @@ class AddExpressionLanguageProvidersPass implements CompilerPassInterface
                 $definition->addMethodCall('registerProvider', [new Reference($id)]);
             }
         }
-
-        if (!$container->hasDefinition('cache.system')) {
-            $container->removeDefinition('cache.security_expression_language');
-            $container->removeDefinition('cache.security_is_granted_attribute_expression_language');
-        }
     }
 }